requirements.txt
asgiref==3.5.2
Django==3.2.15
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
django-js-asset==2.0.0
Pillow==9.2.0
pytz==2022.2.1
sqlparse==0.4.3
